requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.1
pydantic>=2.7.0
# Optional dev tools
black>=23.11.0
flake8>=6.1.0
pytest>=7.4.4
pytest-asyncio>=0.23.6
